i hope am ready to take the challenge.
man am preparing for a challenge in coderbyte for a role a graduate python dev, this is probaly my last push in this repo
i hope when I come back and look at it I will be happy for trying this challenges, I hope this works out. looking forward to making apis with python a td securities.